Split generic arity call site goldens
