+++ tests/integration/cli_build/frontend_json/diagnostics_generic_arity_golden/call_site_annotations/methods.rs
@@ -1,0 +1,85 @@
+use super::super::assert_diagnostics_golden;
+
+#[test]
+fn emit_json_diagnostics_generic_method_type_arg_annotation_arity_schema_matches_golden() {
+    assert_diagnostics_golden(
+        "generic_method_type_arg_annotation_arity.zen",
+        r#"
+Box<T>: {
+    value: T
+}
+
+Holder: {
+    value: i32
+}
+
+Holder.wrap<T> = (self: Holder, value: T) T {
+    value
+}
+
+main = () i32 {
+    holder = Holder { value: 1 }
+    box = Box<i32> { value: 1 }
+    bad = holder.wrap<Box<i32, StaticString>>(box)
+    0
+}
+"#,
+        "generic method type-argument annotation arity",
+        "generic method type-argument annotation arity diagnostics should not emit argument followups",
+        "tests/fixtures/ir_json/diagnostics_generic_method_type_arg_annotation_arity.golden.json",
+    );
+}
+
+#[test]
+fn emit_json_diagnostics_nongeneric_method_type_args_schema_matches_golden() {
+    assert_diagnostics_golden(
+        "nongeneric_method_type_args.zen",
+        r#"
+Box: {
+    value: i32
+}
+
+Box.get = (self: Box) i32 {
+    self.value
+}
+
+main = () i32 {
+    box = Box { value: 1 }
+    box.get<i32>()
+}
+"#,
+        "non-generic method type arguments",
+        "non-generic method type-argument diagnostics should not emit argument followups",
+        "tests/fixtures/ir_json/diagnostics_nongeneric_method_type_args.golden.json",
+    );
+}
+
+#[test]
+fn emit_json_diagnostics_generic_method_type_arg_annotation_missing_args_schema_matches_golden() {
+    assert_diagnostics_golden(
+        "generic_method_type_arg_annotation_missing_args.zen",
+        r#"
+Box<T>: {
+    value: T
+}
+
+Holder: {
+    value: i32
+}
+
+Holder.wrap<T> = (self: Holder, value: T) T {
+    value
+}
+
+main = () i32 {
+    holder = Holder { value: 1 }
+    box = Box<i32> { value: 1 }
+    bad = holder.wrap<Box>(box)
+    0
+}
+"#,
+        "generic method type-argument annotation missing arguments",
+        "generic method type-argument annotation missing-arguments diagnostics should not emit argument followups",
+        "tests/fixtures/ir_json/diagnostics_generic_method_type_arg_annotation_missing_args.golden.json",
+    );
+}
